build(api): Capture Perl command to list nodes
Prototype Perl command to efficiently list nodes recursively.

The most rudimentary API could be just an HTTP static site server to respond with the
node content from the node file, ditto for node attachments.  Of course, it also needs
to be able to create new nodes, and modify and delete existing nodes but similar HTTP
semantics are fine for that we just need verbs/methods other than `GET`.  This is
basically what WebDAV is, so we could use a static HTML server that includes WebDAV
support.  Thankfully [nginx provides a WebDAV
module](https://nginx.org/en/docs/http/ngx_http_dav_module.html).

The purpose of nOrg is supporting arbitrary hierarchies and possibly large ones at that.
As such, the last thing a minimal API needs to provide is efficient recursive listing of
nodes and the VCS system already needs to be efficient at that including knowing
where *not* to walk, so that means the API needs to be able to get the output of `$ git
ls-files`.  Unfortunately, that means the API has just one dynamic requirement beyond a
basic HTTP static site and WebDAV server.  [Nginx still has a Perl
module](https://nginx.org/en/docs/http/ngx_http_perl_module.html#methods) and the
official Docker image has a variant that includes it and it can be used to [return
output from shell commands in the
response](https://stackoverflow.com/a/72199313/624787).  We can also use simple shell
pipelines to sort the listed nodes most recently modified first as that's most likely to
be important to the UI.  It's not the most graceful thing in the world but it works as a
proof of concept and demonstrates minimal requirements without spinning up a separate
project providing a full-fledged API server.  I'm sure Apache has similar options and/or
old-school CGI, but I think they'd have similar kludginess.

This is a working Perl script that will ["stream" nodes one line at a
time](https://stackoverflow.com/a/13402475/624787) as a proof of concept and to try out
different approaches.