build(api): Alternate Perl command to list nodes
I'm not sure which is more efficient when it comes to listing huge numbers of nodes.  I
would guess "streaming" it back line by line as in the previous approach is less
efficient in some ways as it keeps execution in an interpreted language.  OTOH, I'm
guessing this approach is less efficient in other ways including that it loads the full
command output into a string in memory.  OTO*O*H, the `$ sort -n` already has to load
all lines into memory.  So I just wanted to capture both approaches for future
reference.